Address dash shell incompatibility
The 'dash' shell doesn't support the 'pipefail' shell
option. This makes it hard for us to get a correct
exit code when some command in the shell script's
pipe fails. To circumvent this, we completely get
rid of the pipe and redirect to a file - while
combining both stdout and stderr.