BI-800: properly handle power function
MongoDB doesn't handle exponentiation when the base is negative
and the desired root is odd in a manner similar to MySQL. While
MongoDB returns NaN - e.g. : [-4, -4.4] - MySQL returns an
error.
To handle this properly, we don't push down the MySQL pow/power
function and evaluate it in memory instead.