BI-817: Redact password information from logging on mongosqld
This commit restructures the member structs of SqldOptions so that
SqldOptions can be used to log the command line arguments supplied
to mongosqld. This restructuring was necessary because it distinguishes
between flags that are left unspecified and flags that are explicitly
set to false or empty in the command line.