BI-1189: change directory layout of windows release.zip to be the same as *nix
Due to use of DOS paths in a BASH script, we were accidentally
concatenating too many directories in the windows archive (this is a
weird behavior of cygwin bash in general).

make_archive.py will likely not work on windows outside of cywgin, now.
I had to change it to only use '/' ever.  Additionally, it will error
out if a path in a transform contains a backslash