BI-1164: Make MySQL Datetime/Timestamp format consistent
MySQL requires specification of fractional second parts a prior in the
column type.  Since mongo stores the microseconds, it is preferable
that we display them, and return datetime(6) for the type of datetype
columns when "describe table" is called.  This leaves an edge
case in the form of

  select repeat(timestamp('2016-01-01'),2);

returning:

  '2016-01-01 00:00:00.0000002016-01-01 00:00:00.000000'

When it should return:

  '2016-01-01 00:00:002016-01-01 00:00:00'

To fix this we will need a rework of the type system that correctly
disambiguates between values coming out of a collection versus those
that are literals.