BI-1674: Add internal integration tests for TPC-H micro queries
This commit implements changes to the benchmarking infrastructure for BI
Connector such that we are now able to test for correctness of our
benchmark queries. Note that this is a bit more general/expansive than
the ticket requests, since this is a change that should allow _any_
benchmarks to be verified for correctness. In particular however, this
commit only adds such verification for the TPC-H benchmarks. Other
benchmarks would need to have their YAML files modified to also have
their queries be verified, but this is not included in this ticket.

The implementation details can be found in the diff, but in summary,
we add VerifyBenchmarkQuery() and veryBenchmark() functions, as well as
some modifications to the Benchmark struct so that we can read the
expected results from the YAML file and verify them, very much
analogously to how we verify normal test cases.

It is worth noting that this commit also has changes related to floating
point imprecision. This changes our code in algebrizer.go to have
useFloats = true so that certain queries that rely on unreliable
floating point comparisons can pass. This of course results in some
changes to the internal unit tests, as well as the disabling of a
blackbox test (50) that was already unsafe, and discovered to be upon
making this change to floats.