BI-1793, BI-1836: Fix failing GSSAPI cache tests
For test-gssapi-admin-authenticator-plain-user, the initial issue here appeared to be a GLIBC version incompatibility - the C plugin was using RHEL6 while the BIC was using RHEL7.

After upgrading to RHEL7, the version incompatibility issue was resolved - however, the GSSAPI cache behavior had changed so the expected error was changes as well (not a very precise test).

These updates also fixed the test-gssapi-with-correct-username-and-password-wrong-cache failure on OS X.

Summary of Updates
- Run GSSAPI RHEL tests on RHEL7 and not RHEL6
- Update expected errors in tests now running on RHEL7
- Use v1.3.0 of the C plugin for tests.