BI-1853: Null safe equality operator fails on partially, vacuously true multiple term comparisons
This commit fixes an issue where the use of the null safe equality
operator (<=>) causes certain multi-term comparisons to be incorrect
when at least one of the terms are vacuously true.

e.g.
  SELECT * FROM foo WHERE (d, d) <=> (d, a);

Will return all rows in `foo`, instead of only returning rows that
satisfy the propert that (d, d) is null-safe-equals to (d, a) (or
equivalently, that d is null-safe-equal to a).

The change for this is actually only the addition of two lines to
`translateTupleExpr()` in `algebrizer.go`.