BI-1222: Fix SCRAM-SHA test failure
This is a small change to remove the previously failing
test-auth-full-schema-available test. This test was considered to not
provide much value, yet take away from developer productivity, and was
therefore removed. With this change, BI-1222's changes (of adding a new
test database) will no longer cause failures on Evergreen.