BI-1785: In-memory arithmetic failure
This commit fixes up issues with our arithmetic. In particular, this
commit reverts the `useFloats = true` change introduced in BI-1674, and
then accordingly keeps the TPC-H benchmarks passing by
updating/correcting the `mongod` dataset being used by the benchmarks
(this change is not included in the commit as it is not tracked under
git). Part of this reversion also includes updating the pre-existing
tests that were changed to accomodate floats back to using Decimal128.

Finally, we also correct an issue within tpch.yml, where floating point
values were not written inside the query as 0.05 and 0.07 as they
should, but actually things like 0.05000...003. This was probably
introduced by a previousl developer as a way to circumvent previous
floating point issues.

On top of this, we also added back select_60, as this test now passes
with these changes.