BI-1891: Fix increase of timeout of benchmark-queries
This commit is a single line addition to our .evg.yml file that includes
an exec_timeout_secs parameter set to 8 hours for benchmark-queries-3.4.
This parameter value was forgotten in BI-1851, which was causing some
inconsistent timeout behavior from the related task. This problem should
now be fixed, as seen in the patch builds here:

Standard PS1 running benchmark-queries-3.4:
https://evergreen.mongodb.com/version/5bbcf8aa2a60ed60ba24b012

Ensuring that timeout parameter is honored (set to 1 minute):
https://evergreen.mongodb.com/version/5bbcf5a92fbabe755484a453

Artifcially slowing the benchmarks with sleep calls and ensuring that we
time it out at 8 hours:
https://evergreen.mongodb.com/version/5bbe2fdd2fbabe5e09092bdf