BI-2087: INSERT() fails to replace substring when pushed down (#60)
Since the MDB aggregation language starts its counting from 0, and MySQL starts its counting  from 1, we do a subtraction of 1 to "translate" the MySQL counting over to MDB aggregation. However, this means that our checks for an invalid position should check for something being _less than_ 0, *not* _less than or equal to_ 0, since <= 0 is actually valid and refers to when the MySQL pos is 1. This commit fixes this by changing the aggregation operation to `OpLt` instead of `OpLte` and updates the corresponding tests, and then adds a test for this pos = 1 case.