BI-2137: Add config option to allow constrained delegation (#100)
Rather than unconditionally allowing constrained delegation, which lead to keytab configuration issues for those not using constrained delegation with a linux mongosqld, we require the flag be set of constrained delegation is desired. This keeps older configurations from breaking, while allowing constrained delegation to work with a linux mongosqld (it previously would work with a Windows mongosqld).