SQL-2298: fix test_panic_safe_exec (#809)
Wrapping in a once so that we only set the hook one time. I ran a patch on evergreen with 10 duplicates of the commented test to ensure we had a lot of tests panic at the same time, and it seems to have resolved the issue.

Additionally, I removed the other `set_hook` since we never did anything with it.