SQL-2335: add misc search stages (#829)
* SQL-2335: add misc search stages

According to the technical design, $search, $searchMeta, and $vectorSearch are no-op stages, so I aliased them all to unsupported stages. I did add a test for each to ensure they parse correctly, and I exteded air agg_ast in the same manner we did for expressions we don't plan to support there.

* SQL-2335: get variant of AtlasSearchStage

* feedback